userLocale export to allow for custom plugins to use it
