+++ lib/ui/utils/_utils.mjs
@@ -55,6 +55,7 @@ import jsonDataview from './jsonDataview.mjs';
 
 import locationCount from './locationCount.mjs';
 import resizeHandler from './resizeHandler.mjs';
+import userLocale from './userLocale.mjs';
 
 export default {
   Chart,
@@ -66,4 +67,5 @@ export default {
   locationCount,
   resizeHandler,
   Tabulator,
+  userLocale,
 };
